refactor: add abilities runtime workspace skeleton
