+++ src-tauri/abilities-runtime/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "abilities-runtime"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+async-trait = "0.1"
+base64 = "0.22"
+chrono = { version = "0.4", features = ["serde"] }
+hex = "0.4"
+http = "1"
+inventory = "0.3"
+parking_lot = { version = "0.12", features = ["arc_lock"] }
+rand = "0.10"
+schemars = "0.8.22"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+sha2 = "0.10"
+thiserror = "2"
+tracing = "0.1"
+url = "2"
+uuid = { version = "1", features = ["v4", "serde"] }
+
+[dev-dependencies]
+tokio = { version = "1", features = ["macros", "rt", "time"] }
+
+[features]
+experimental = []
+harness-hermetic = []
