+++ src-tauri/abilities-runtime/src/lib.rs
@@ -1,0 +1,17 @@
+#![forbid(unsafe_code)]
+
+pub mod intelligence {
+    pub mod provider;
+}
+pub mod observability;
+pub mod sensitivity;
+pub mod services {
+    pub mod context;
+    pub mod external_replay;
+    pub mod sensitivity {
+        pub use crate::sensitivity::*;
+    }
+}
+pub mod types;
+
+pub use intelligence::provider::*;
