+++ src-tauri/src/services/sensitivity.rs
@@ -12,160 +12,13 @@ use crate::intelligence::{
     CompanyContext, CurrentState, IntelRisk, IntelWin, IntelligenceJson, StakeholderInsight,
     ValueItem,
 };
+pub use abilities_runtime::sensitivity::{
+    RedactionAffordance, RenderActor, RenderDecision, RenderPolicy, RenderPolicyKind,
+    RenderSurface, RenderableClaimText,
+};
 use chrono::Utc;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-pub enum RenderSurface {
-    TauriEntityDetail,
-    TauriBriefingPrep,
-    TauriMeetingDetail,
-    TauriEmailSummary,
-    TauriProvenance,
-    TauriReport,
-    TauriChat,
-    McpTool,
-    McpToolDetail,
-    P2Publication,
-    LogStructured,
-    PushNotification,
-}
-
-impl RenderSurface {
-    pub fn from_name(value: &str) -> Option<Self> {
-        match value.trim().to_ascii_lowercase().as_str() {
-            "tauri_entity_detail" | "entity_detail" => Some(Self::TauriEntityDetail),
-            "tauri_briefing_prep" | "briefing_prep" => Some(Self::TauriBriefingPrep),
-            "tauri_meeting_detail" | "meeting_detail" => Some(Self::TauriMeetingDetail),
-            "tauri_email_summary" | "email_summary" => Some(Self::TauriEmailSummary),
-            "tauri_provenance" | "provenance" => Some(Self::TauriProvenance),
-            "tauri_report" | "report" => Some(Self::TauriReport),
-            "tauri_chat" | "chat" => Some(Self::TauriChat),
-            "mcp_tool" => Some(Self::McpTool),
-            "mcp_tool_detail" => Some(Self::McpToolDetail),
-            "p2_publication" => Some(Self::P2Publication),
-            "log_structured" => Some(Self::LogStructured),
-            "push_notification" => Some(Self::PushNotification),
-            _ => None,
-        }
-    }
-
-    fn allows_reveal(self) -> bool {
-        matches!(
-            self,
-            Self::TauriEntityDetail
-                | Self::TauriMeetingDetail
-                | Self::TauriEmailSummary
-                | Self::TauriProvenance
-                | Self::TauriReport
-        )
-    }
-
-    fn is_first_party_tauri(self) -> bool {
-        matches!(
-            self,
-            Self::TauriEntityDetail
-                | Self::TauriBriefingPrep
-                | Self::TauriMeetingDetail
-                | Self::TauriEmailSummary
-                | Self::TauriProvenance
-                | Self::TauriReport
-        )
-    }
-
-    fn is_agent_surface(self) -> bool {
-        matches!(self, Self::McpTool | Self::McpToolDetail | Self::TauriChat)
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct RenderActor {
-    pub actor: String,
-    pub user_id: Option<String>,
-}
-
-impl RenderActor {
-    pub fn user(actor: impl Into<String>, user_id: Option<impl Into<String>>) -> Self {
-        Self {
-            actor: actor.into(),
-            user_id: user_id.map(Into::into),
-        }
-    }
-
-    pub fn agent(actor: impl Into<String>) -> Self {
-        Self {
-            actor: actor.into(),
-            user_id: None,
-        }
-    }
-
-    fn is_user(&self) -> bool {
-        self.actor.trim().eq_ignore_ascii_case("user")
-            || self.actor.trim().to_ascii_lowercase().starts_with("user:")
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, JsonSchema)]
-#[serde(rename_all = "snake_case", tag = "kind")]
-pub enum RedactionAffordance {
-    ConfidentialClickToReveal {
-        claim_id: String,
-        label: String,
-        audit_required: bool,
-    },
-    ConfidentialHidden {
-        label: String,
-    },
-    UserOnlyHidden {
-        label: String,
-    },
-}
-
-impl RedactionAffordance {
-    pub fn label(&self) -> &str {
-        match self {
-            Self::ConfidentialClickToReveal { label, .. }
-            | Self::ConfidentialHidden { label }
-            | Self::UserOnlyHidden { label } => label,
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-pub enum RenderPolicyKind {
-    Render,
-    Redacted,
-    Drop,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, JsonSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct RenderPolicy {
-    pub kind: RenderPolicyKind,
-    pub sensitivity: ClaimSensitivity,
-    pub surface: RenderSurface,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub claim_id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub affordance: Option<RedactionAffordance>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum RenderDecision {
-    Render,
-    RenderRedacted { affordance: RedactionAffordance },
-    Drop,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, JsonSchema)]
-#[serde(rename_all = "camelCase")]
-pub struct RenderableClaimText {
-    pub text: String,
-    pub policy: RenderPolicy,
-}
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, JsonSchema)]
 #[serde(rename_all = "snake_case")]
